+++ src/ifs/wl_seat/tablet/zwp_tablet_manager_v2.rs
@@ -57,7 +57,7 @@ impl Global for ZwpTabletManagerV2Global {
     }
 
     fn version(&self) -> u32 {
-        1
+        2
     }
 }